# Backend/internal/docs/api.yaml
openapi: 3.0.3
info:
  title: ALX Real Estate API
  description: REST API for ALX Real Estate — an Airbnb-style property marketplace
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
    description: Local development
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: resource not found
        detail:
          type: string
          example: additional context
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        full_name:
          type: string
        role:
          type: string
          enum: [guest, host, admin]
        avatar_url:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
          description: JWT valid for 15 minutes. Send as Authorization Bearer token on protected routes.
        user:
          $ref: '#/components/schemas/User'
...
paths:
...
  /auth/refresh:
    post:
      summary: Refresh access token
      tags: [Auth]
      description: >
        Uses the `refresh_token` HttpOnly cookie set during login/register.
        No request body needed — the browser sends the cookie automatically.
        Returns a new access token and rotates the refresh token cookie.
      responses:
        '200':
          description: New access token issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Missing or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...